Move unpack fix (#395)
Elo | 0.83 +- 2.52 (95%)
SPRT | 8.0+0.08s Threads=1 Hash=16MB
LLR | 2.94 (-2.94, 2.94) [-3.00, 1.00]
Games | N: 34140 W: 8030 L: 7948 D: 18162
Penta | [129, 4067, 8599, 4143, 132]

Bench: 6373337

+++ src/random.h
@@ -3,10 +3,10 @@
 #include "types.h"
 
 // pseudo random number state
-extern unsigned int random_state;
+extern uint64_t random_state;
 
 // generate 32-bit pseudo legal numbers
-[[nodiscard]] unsigned int GetRandomU32Number();
+[[nodiscard]] uint64_t GetRandomU32Number();
 
 // generate 64-bit pseudo legal numbers
 [[nodiscard]] Bitboard GetRandomBitboardNumber();
